cleanup plotting utilities tests
